# Курсовая/CursovoiProject/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Veterinary Clinic management system.",
        "title": "Veterinary Clinic API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/appointments/{id}/complete": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Закрывает визит и сохраняет медицинский протокол (диагноз, вес, лечение)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Завершение приема",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные осмотра",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CompleteVisitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/appointments/{id}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Принять (confirmed) или отклонить (rejected) запись",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Изменить статус записи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
...
        "/doctor/patients": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search for clients by name or phone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Search patients",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PatientDTO"
                            }
                        }
                    }
                }
            }
        },
        "/doctor/patients/{id}/history": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получает историю всех визитов всех питомцев конкретного клиента",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Медицинская история клиента",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PetHistoryDTO"
                            }
                        }
                    }
                }
            }
        },
        "/doctor/schedule": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Список всех записей к конкретному врачу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doctor"
                ],
                "summary": "Расписание врача",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID врача",
                        "name": "doctor_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.DoctorScheduleDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/pets": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Добавить питомца",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID владельца",
                        "name": "ownerId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Данные питомца",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.PetResponse"
                        }
                    }
                }
            }
        },
        "/pets/owner/{userId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Список питомцев владельца",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID владельца",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PetResponse"
                            }
                        }
                    }
                }
            }
        },
        "/pets/{petId}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Редактировать питомца",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID питомца",
                        "name": "petId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PetResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "pets"
                ],
                "summary": "Удалить питомца",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID питомца",
                        "name": "petId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CompleteVisitRequest": {
            "type": "object",
            "required": [
                "diagnosis_ids",
                "weight"
            ],
            "properties": {
                "diagnosis_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "medications": {
                    "type": "string"
                },
                "treatment": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
        "dto.DiagnosisDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.DoctorScheduleDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ownerName": {
                    "type": "string"
                },
                "petName": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "dto.PatientDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastVisit": {
                    "type": "string"
                },
                "petsCount": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "dto.PetHistoryDTO": {
            "type": "object",
            "properties": {
                "petIcon": {
                    "type": "string"
                },
                "petName": {
                    "type": "string"
                },
                "visits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.VisitTimelineDTO"
                    }
                }
            }
        },
        "dto.PetRequest": {
            "type": "object",
            "required": [
                "dob",
                "name",
                "species"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "species": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "dto.PetResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "petId": {
                    "type": "integer"
                },
                "species": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
        "dto.VisitTimelineDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "diagnoses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DiagnosisDTO"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "treatment": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}